+++ Runtime.h
@@ -357,8 +357,8 @@ namespace Runtime
     uint64_t tlsAlloc();
     void     tlsSetValue(uint64_t id, void* value);
     void*    tlsGetValue(uint64_t id);
-    void*    alloc(uint32_t size);
-    void*    realloc(void* ptr, uint32_t size);
+    void*    alloc(size_t size);
+    void*    realloc(void* ptr, size_t size);
     void     free(void* ptr);
 
     const uint32_t COMPARE_STRICT   = 0x00000000;
